% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat_barra.R
\name{estat_barra}
\alias{estat_barra}
\title{Gráfico de barras no estilo Estat Júnior}
\usage{
estat_barra(
  dados,
  grupos,
  fill_by = NULL,
  wrap_by = NULL,
  paleta = NULL,
  rotulo = NULL
)
}
\arguments{
\item{dados}{(obrigatório) dataframe ou tibble. O banco de dados com as variáveis de interesse para gerar o gráfico.}

\item{grupos}{(obrigatório) vetor de caracteres. Define as variáveis a serem utilizadas para agrupar e calcular as frequências.
Essa função não suporta vetores com mais de 2 variáveis pois plotar mais que isso em um gráficos de barras
signigfica perder a interpretabilidade.}

\item{fill_by}{character. Quando trabalhamos com 2 grupos, precisamos escolher alguma forma de inserir
a segunda variável no gráfico. Esse parâmetro define qual das variáveis vai aparecer colorindo as barras de acordo
com suas categorias.}

\item{wrap_by}{character. Quando trabalhamos com 2 grupos, precisamos escolher alguma forma de inserir
a segunda variável no gráfico. Esse parâmetro define qual das variáveis vai aparecer separando os gráficos
em painéis conforme suas categorias.}

\item{paleta}{caractere/vetor. Escolha cores para utilizar no gráfico. Tenha em mente que o número de cores
especificadas deve ser o mesmo do número de categorias da variável.}

\item{rotulo}{caractere. Escolha um método para adicionar um rótulo com a frequência das categorias.
O parâmetro reconhece os valores: \code{"n"}, para frequência absoluta, \code{"p"},
para frequência em porcentagem e \code{"np"} ou \code{"pn"} para frequência absoluta e em porcentagem.}
}
\value{
o gráfico conforme configurado.
}
\description{
Cria gráficos de barras com os grupos, frequências, métodos, temas and outras características
de acordo com o padrão utilizado pela Estat para gráficos de barra.
}
